+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for NeoFS Node
 - Erasure coding in containers' storage policies (#3570)
 - IR `experimental.allow_ec` config option (#3570)
 - SN `pprof.enable_block` and `pprof.enable_mutex` options (#3655)
+- `neofs-adm fschain load-report` command (#3649)
 
 ### Fixed
 - Write cache using too much CPU (#3642)